Updated the Email Validation
Currently Email validation works as long as there is alpha numeric before and after the '@' symbol and after the fullstop.

However, failed email login will still be added to the DB. Next update will include moving all middle ware to their appropriate area/ disallow the emails to be written to the DB if they are of the wrong format

need front end to implement a redirect/refresh of the page to allow user to input username and password again.